+++ src/Contracts/Jsonable.php
@@ -1,0 +1,7 @@
+<?php
+namespace Budgetlens\CopernicaRestApi\Contracts;
+
+interface Jsonable
+{
+    public function toJson(int $options = 0): string;
+}