+++ src/Exceptions/CopernicaApiException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Budgetlens\CopernicaRestApi\Exceptions;
+
+class CopernicaApiException extends \Exception
+{
+}