+++ src/Resources/Account/Identity.php
@@ -1,0 +1,12 @@
+<?php
+namespace Budgetlens\CopernicaRestApi\Resources\Account;
+
+use Budgetlens\CopernicaRestApi\Resources\BaseResource;
+
+class Identity extends BaseResource
+{
+    public $id;
+    public $name;
+    public $description;
+    public $company;
+}