+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 All notable changes to `copernica-rest-api-client` will be documented in this file.
 
 ## [Unreleased]
-- 
+- Account endpoint
+  - identity
+  - consumption